+++ src/test/java/net/kccricket/bestesttool/commands/CommandRefillTest.java
@@ -59,9 +59,8 @@ class CommandRefillTest extends BestToolsTestBase {
 
     @Test
     void nonPlayerSenderIsRejected() {
-        server.dispatchCommand(server.getConsoleSender(), "refill");
-
-        assertEquals("You must be a player to run this command.", server.getConsoleSender().nextMessage());
+        assertLogMessageContains("You must be a player to run this command.",
+                () -> server.dispatchCommand(server.getConsoleSender(), "refill"));
     }
 
     @Test